requirements.txt
dlib
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.2.9
h5py==2.10.0
joblib==0.14.1
jsonschema==3.2.0
Keras==2.3.1
matplotlib==3.2.1
mtcnn==0.1.0
numpy==1.18.4
opencv-python==4.2.0.34
paho-mqtt==1.5.0
pandas==1.0.3
Pillow==7.1.2
requests==2.22.0
scikit-image==0.17.1
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.1
sklearn==0.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
torch==1.5.0
torchvision==0.6.0
tqdm==4.46.0